+++ lib/questionnaires/model/item/answer/src/answer_model.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import 'package:fhir/r4.dart';
 import 'package:flutter/material.dart';
 
@@ -10,8 +8,6 @@ import '../../../model.dart';
 abstract class AnswerModel<I, V> extends ResponseNode {
   /// Textual depiction of an unanswered question.
   static const nullText = '—';
-
-  static const _htmlEscaper = HtmlEscape();
 
   final QuestionItemModel responseItemModel;
   V? _value;
@@ -57,13 +53,8 @@ abstract class AnswerModel<I, V> extends ResponseNode {
 
   /// Returns a human-readable, localized, textual description of the model.
   ///
-  /// Returns [nullText] if the question is unanswered.
-  String get display;
-
-  /// Same as [display], but safely escaped for HTML/XHTML.
-  String get xhtmlDisplay {
-    return AnswerModel._htmlEscaper.convert(display);
-  }
+  /// Returns [XhtmlString.nullText] if the question is unanswered.
+  XhtmlString get display;
 
   /// Validates a new input value. Does not change the [value].
   ///